Ensure spec files are up to date before executing (#2)
Whenever there's a bug or unclear test message that we want to update after students have already started we can't affect students spec files.

This commit enhances rails grade check if the local spec folder/files are up to date with the latest version stored on appdev-projects/upstream version. If they do match, the normal rails grade thing happens. If they don't match, any new files in the spec folder are discarded, the updated files are downloaded, the specs are run, and the results are finally submitted to Grades.